feat: Add default retention rate to suppliers
Introduces a `defaultRetentionRate` field to the `Supplier` type. This allows storing a default retention percentage (75% or 100%) for suppliers.

When an invoice is analyzed, if the identified supplier has a `defaultRetentionRate` set, this value will now be used to pre-fill the retention percentage in the invoice wizard, streamlining the data entry process.

Also updates the `App.tsx` state management to include `editingSupplier` and `isSavingSupplier` states for future supplier editing functionality.